+++ src/gpu/intel/jit/ir/linear_expr.hpp
@@ -50,7 +50,7 @@ expr_t linear_div(const expr_t &e, int64_t factor);
 
 // Simplifies a modulus of an expression by a constant. The expression must be
 // convertable to linear_t.
-expr_t simplify_linear_mod(const expr_t &e, int factor);
+expr_t simplify_linear_mod(const expr_t &e, int64_t factor);
 
 // Returns the base and the increments of linear expression `expr` when
 // incrementing `idxs[i]` by 1:
